# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BankManager API",
        "description": "API de gestion bancaire pour BankManager",
        "contact": {
            "email": "contact@bankmanager.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur de développement"
        }
    ],
    "paths": {
...
        "/api/transactions": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Lister toutes les transactions",
                "description": "Récupère la liste paginée de toutes les transactions",
                "operationId": "9b50e29e364179ebde24ac5fc59cd442",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de la page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Nombre d'éléments par page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "compte_id",
                        "in": "query",
                        "description": "Filtrer par compte UUID",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Filtrer par type de transaction",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "depot",
                                "retrait",
                                "virement_entrant",
                                "virement_sortant",
                                "frais",
                                "interet"
                            ]
                        }
                    },
                    {
                        "name": "date_debut",
                        "in": "query",
                        "description": "Date de début (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "date_fin",
                        "in": "query",
                        "description": "Date de fin (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des transactions récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Transaction"
                                            }
                                        },
                                        "links": {
                                            "$ref": "#/components/schemas/PaginationLinks"
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "passport": [
                            "read-transactions"
                        ]
                    }
                ]
            },
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Créer une nouvelle transaction",
                "description": "Effectue une nouvelle transaction sur un compte bancaire",
                "operationId": "54eaa3952808f639d46f969f4c8b98c5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "compte_id",
                                    "type",
                                    "montant"
                                ],
                                "properties": {
                                    "compte_id": {
                                        "description": "UUID du compte",
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "type": {
                                        "description": "Type de transaction",
                                        "type": "string",
                                        "enum": [
                                            "depot",
                                            "retrait",
                                            "virement_entrant",
                                            "virement_sortant",
                                            "frais",
                                            "interet"
                                        ]
                                    },
                                    "montant": {
                                        "description": "Montant de la transaction",
                                        "type": "number",
                                        "format": "decimal"
                                    },
                                    "description": {
                                        "description": "Description de la transaction",
                                        "type": "string"
                                    },
                                    "compte_destinataire": {
                                        "description": "Numéro du compte destinataire (pour virements)",
                                        "type": "string"
                                    },
                                    "nom_destinataire": {
                                        "description": "Nom du destinataire (pour virements)",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Transaction créée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Transaction"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Données invalides ou solde insuffisant"
                    },
                    "404": {
                        "description": "Compte non trouvé"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "passport": [
                            "write-transactions"
                        ]
                    }
                ]
            }
        },
        "/api/transactions/{id}": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Afficher une transaction",
                "description": "Récupère les détails d'une transaction spécifique",
                "operationId": "45e35076645e40d84080bf253c06a689",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID de la transaction",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction trouvée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Transaction"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Transaction non trouvée"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "passport": [
                            "read-transactions"
                        ]
                    }
                ]
            },
            "put": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Annuler une transaction",
                "description": "Annule une transaction (si elle n'est pas déjà validée)",
                "operationId": "43fb5702f8299f2d383e663429920d7a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID de la transaction",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction annulée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Transaction"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Transaction ne peut pas être annulée"
                    },
                    "404": {
                        "description": "Transaction non trouvée"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "passport": [
                            "write-transactions"
                        ]
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Supprimer une transaction",
                "description": "Supprime une transaction (réservé aux administrateurs)",
                "operationId": "bc63b06ba9e12cac8ef887c638602dd5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID de la transaction",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Transaction supprimée avec succès"
                    },
                    "404": {
                        "description": "Transaction non trouvée"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "passport": [
                            "write-transactions"
                        ]
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Transaction": {
                "title": "Transaction",
                "description": "Modèle représentant une transaction bancaire",
                "properties": {
                    "id": {
                        "description": "UUID de la transaction",
                        "type": "string",
                        "format": "uuid"
                    },
                    "type": {
                        "description": "Type de transaction",
                        "type": "string",
                        "enum": [
                            "depot",
                            "retrait",
                            "virement_entrant",
                            "virement_sortant",
                            "frais",
                            "interet"
                        ]
                    },
                    "type_libelle": {
                        "description": "Libellé du type de transaction",
                        "type": "string"
                    },
                    "montant": {
                        "description": "Montant de la transaction",
                        "type": "number",
                        "format": "decimal"
                    },
                    "montant_formate": {
                        "description": "Montant formaté avec devise",
                        "type": "string"
                    },
                    "devise": {
                        "description": "Devise de la transaction",
                        "type": "string",
                        "maxLength": 3
                    },
                    "description": {
                        "description": "Description de la transaction",
                        "type": "string",
                        "nullable": true
                    },
                    "date_transaction": {
                        "description": "Date et heure de la transaction",
                        "type": "string",
                        "format": "datetime"
                    },
                    "statut": {
                        "description": "Statut de la transaction",
                        "type": "string",
                        "enum": [
                            "en_attente",
                            "validee",
                            "annulee",
                            "echouee"
                        ]
                    },
                    "compte_destinataire": {
                        "description": "Numéro du compte destinataire",
                        "type": "string",
                        "nullable": true
                    },
                    "nom_destinataire": {
                        "description": "Nom du destinataire",
                        "type": "string",
                        "nullable": true
                    },
                    "frais": {
                        "description": "Frais de transaction",
                        "type": "number",
                        "format": "decimal"
                    },
                    "solde_apres": {
                        "description": "Solde après transaction",
                        "type": "number",
                        "format": "decimal"
                    },
                    "created_at": {
                        "description": "Date de création",
                        "type": "string",
                        "format": "datetime"
                    },
                    "updated_at": {
                        "description": "Date de mise à jour",
                        "type": "string",
                        "format": "datetime"
                    },
                    "compte": {
                        "$ref": "#/components/schemas/Compte"
                    }
                },
                "type": "object"
            },
...
            "PaginationLinks": {
                "title": "PaginationLinks",
                "description": "Liens de pagination",
                "properties": {
                    "first": {
                        "description": "Lien vers la première page",
                        "type": "string"
                    },
                    "last": {
                        "description": "Lien vers la dernière page",
                        "type": "string"
                    },
                    "prev": {
                        "description": "Lien vers la page précédente",
                        "type": "string",
                        "nullable": true
                    },
                    "next": {
                        "description": "Lien vers la page suivante",
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "PaginationMeta": {
                "title": "PaginationMeta",
                "description": "Métadonnées de pagination",
                "properties": {
                    "current_page": {
                        "description": "Page actuelle",
                        "type": "integer"
                    },
                    "from": {
                        "description": "Premier élément de la page",
                        "type": "integer"
                    },
                    "last_page": {
                        "description": "Dernière page",
                        "type": "integer"
                    },
                    "links": {
                        "description": "Liens de pagination",
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    },
                    "path": {
                        "description": "URL de base",
                        "type": "string"
                    },
                    "per_page": {
                        "description": "Éléments par page",
                        "type": "integer"
                    },
                    "to": {
                        "description": "Dernier élément de la page",
                        "type": "integer"
                    },
                    "total": {
                        "description": "Nombre total d'éléments",
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel Passport OAuth2 security.",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost:8000/oauth/authorize",
                        "tokenUrl": "http://localhost:8000/oauth/token",
                        "refreshUrl": "http://localhost:8000/oauth/token/refresh",
                        "scopes": {
                            "read-clients": "Lire les clients",
                            "write-clients": "Créer/modifier les clients",
                            "read-comptes": "Lire les comptes",
                            "write-comptes": "Créer/modifier les comptes",
                            "read-transactions": "Lire les transactions",
                            "write-transactions": "Créer des transactions"
                        }
                    }
                }
            }
        }
    },
...
    "security": [
        {
            "passport": [
                "read-clients",
                "write-clients",
                "read-comptes",
                "write-comptes",
                "read-transactions",
                "write-transactions"
            ]
        }
    ]
}